Add ConfigMap tracking tests
